requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.8.0
tqdm>=4.66.0
pyyaml>=6.0.1
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tensorboard>=2.15.0
flask>=2.3.0
# Optional: for faster video decoding
decord>=0.6.0 